Update 1.20 | Tools and WA Calculator
This is major update while it isn't really advanced yet I hope it will in future. That's why this update is 1.20 not 2.0 yet.

Added:
- Tools main page
- Tools weighted average calculator page
- scripts behind weighted average calculator.
- CSS files so the calculator works.
- Changed meta so the image is better
- Configs in Weighted Average Calculator, You can set weights and grades and then save the config which will be saved in localStorage.
- Of course added features in tools main and calculator page, like custom css.
- added meta tags into tools and weighted average calculator pages.

This may seem like nothing but it's actually more. The script is around 450 lines long. And the css is also big, Well check it out: https://vulpidev.github.io/tools/index